use endian agnostic file reader
this is a ByteArray class that mimics the ActionScript class of the same
name. it was written for my FLOD modplayer library conversion.
it is hereby released free for usage by AGS.

basically the old code was completely bogus and any int read on a big
endian machine would be wrong.

by using this, we always get the correct value.