fix UB when calling variadic functions from scripts
using a variadic function with a non-variadic function pointer is UB
and crashes on AMD64 due to calling conventions.

with this fix, it is actually possible now to play a couple of minutes
when one is lucky. H2G2 crashed almost immediately before this fix
when compiled with GCC 4.5.1.

it is not stable now, what needs to be done in order to get all function
pointer related bugs fixed is create launcher functions for each
function type used.
i.e. for a function
Object* xxx (Object*, int, char*);
there must be a corresponding wrapper that calls a
void* (void*, int, void*); function pointer.
it is probably sufficient when the sizes of the return value/arguments match.

additionally my way of checking the function name is very inefficient,
one should add a way to quickly check a functions args and return value
and use the appropriate wrapper function.